fix: Fix incorrect splash zone evaluation
We were only evaluating the bottom and lower arc of points in the splash
zone for overlap with the map polygon, and ignoring the interior of the
splash zone.

This is obviously incorrect since the entire zone needs to be inside the
map. This lead to the pathological behavior we saw with group 6's snake
map, where the bottom and top arc were in the map, but the majority of
the splash zone actually overlapped with the water.

NOTE: Shapely's calculation for polygon intersection is faster than
sympy, but it's still slower than what we were doing with matplotlib.
We should look into finding a faster algorithm, or possibly writing one
using numba to increase perf.